Add Dockerfile, Mock recorder, Free space reporting, recordOrder cancellation
+ added dockerfile
+ created mock recorder for development while other instance is running
+ Added detecting free space
# recordOrders deleted on server will be cancelled

+++ tuner/src/main/java/com/tuner/model/server_requests/HeatbeatRequest.java
@@ -6,12 +6,12 @@ import lombok.Data;
 public class HeatbeatRequest {
 
     String id;
-    int freeSpace;
+    long freeSpace;
     boolean isRecording;
     int currentRecordingTime;
     int currentRecordingSize;
 
-    public HeatbeatRequest(String id, int freeSpace, boolean isRecording, int currentRecordingTime, int currentRecordingSize) {
+    public HeatbeatRequest(String id, long freeSpace, boolean isRecording, int currentRecordingTime, int currentRecordingSize) {
         this.id = id;
         this.freeSpace = freeSpace;
         this.isRecording = isRecording;
